backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pymysql>=1.1.0
